Order placed email, dashboard tweaks.

+++ resources/views/dashboard/leads/index.blade.php
@@ -1,22 +1,43 @@
-<x-app-layout>
+@php
+    $breadcrumbs = [
+        [
+            'name' => 'Dashboard',
+            'link' => route('dashboard'),
+        ],
+        [
+            'name' => 'Leads',
+            'link' => false,
+        ],
+    ];
+    $title = 'Contact Leads';
+@endphp
+
+@extends('layouts.app', [
+    'title' => $title,
+    'breadcrumbs' => $breadcrumbs,
+    'addButton' => true,
+    'btn' => [
+        'text' => 'Back',
+        'link' => route('dashboard'),
+    ],
+])
+
+@section('title', $title)
+
+@push('styles')
     <style>
         .delete {
             height: 40px;
             width: 40px;
             border-radius: 50%;
-            color: white
         }
-
-        .delete:hover {
-            color: white
+        .swal2-popup .swal2-styled {
+            margin: 0 10px !important;
         }
     </style>
-    <x-slot name="header">
-        <h2 class="font-semibold text-xl text-gray-800 leading-tight">
-            {{ __('Contact Leads') }}
-        </h2>
-    </x-slot>
+@endpush
 
+@section('content')
     <div class="container-fluid">
         <div class="p-5">
             <div class="table-responsive">
@@ -40,9 +61,8 @@
                                     <td>{{ $lead->message }}</td>
                                     <td>
                                         <a class="delete bg-danger d-flex align-items-center justify-content-center"
-                                            href="javascript:void(0)"
-                                            data-url="{{ route('contact.destroy', $lead->id) }}">
-                                            <i class="fas fa-trash-alt"></i>
+                                            href="javascript:void(0)" data-url="{{ route('contact.destroy', $lead->id) }}">
+                                            <i class="fas fa-trash-alt text-white"></i>
                                         </a>
                                     </td>
                                 </tr>
@@ -61,32 +81,42 @@
             </div>
         </div>
     </div>
+@endsection
 
-    @section('scripts')
-        <script>
-            $('.delete').on('click', function() {
-                axios.delete($(this).data('url'))
-                    .then(response => response.data)
-                    .then(data => {
-                        if (data.success) {
-                            window.location.reload()
-                        } else {
+@push('scripts')
+    <script>
+        $('.delete').on('click', function() {
+            swal.fire({
+                icon: 'warning',
+                title: 'Confirm delete',
+                text: 'Are you sure to proceed?',
+                showCancelButton: true,
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    axios.delete($(this).data('url'))
+                        .then(response => response.data)
+                        .then(data => {
+                            if (data.success) {
+                                window.location.reload()
+                            } else {
+                                swal.fire({
+                                    icon: 'error',
+                                    title: 'Oops...',
+                                    text: 'Something went wrong!',
+                                })
+                            }
+                        })
+                        .catch(err => {
                             swal.fire({
                                 icon: 'error',
                                 title: 'Oops...',
                                 text: 'Something went wrong!',
                             })
-                        }
-                    })
-                    .catch(err => {
-                        swal.fire({
-                            icon: 'error',
-                            title: 'Oops...',
-                            text: 'Something went wrong!',
                         })
-                    })
+                }
             })
-        </script>
-    @endsection
 
-</x-app-layout>
+
+        })
+    </script>
+@endpush
